feat(store): Removing unused stuff
