+++ src/assets/javascripts/index.js
@@ -2,10 +2,10 @@ import Rx from 'rx';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import gameStore from './store/game-store';
+import stateStore from './store/state-store';
 import App from './components/app';
 
-gameStore.subscribe((store) => {
-  console.log('from component', store);
+stateStore.subscribe((store) => {
+  console.log('component update store: ', store);
   // ReactDOM.render(<App store={store} />, document.getElementById('app'));
 });
